% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_output_subfuncs.R
\name{ca_ms_anom_nt}
\alias{ca_ms_anom_nt}
\title{Cohort Attrition \emph{Multi Site Anomaly No Time}}
\usage{
ca_ms_anom_nt(process_output, output)
}
\arguments{
\item{process_output}{output from compute_attrition_diff}

\item{output}{the column that should be used as the y-axis:

\if{html}{\out{<div class="sourceCode">}}\preformatted{          options are:
          `num_pts` (raw patient count),
          `prop_retained_start` (proportion patients retained from starting step),
          `prop_retained_prior` (proportion patients retained from prior step),
          `prop_diff_prior` (proportion difference between each step and the prior step)
}\if{html}{\out{</div>}}}
}
\value{
a dot plot indicating anomalous sites based on the user selected output value

\if{html}{\out{<div class="sourceCode">}}\preformatted{    anomalies are indicated by STARS, the color of each dot represents the raw output value,
    and the size of each dot represents the mean output value per attrition step
}\if{html}{\out{</div>}}
}
\description{
Cohort Attrition \emph{Multi Site Anomaly No Time}
}
