% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_output_subfuncs.R
\name{ca_ms_exp_nt}
\alias{ca_ms_exp_nt}
\title{Cohort Attrition \emph{Multi Site Exploratory No Time}}
\usage{
ca_ms_exp_nt(process_output, log_scale = FALSE, output = "num_pts")
}
\arguments{
\item{process_output}{output from compute_attrition_diff}

\item{log_scale}{logical to determine whether a log transform should be applied
to the y axis}

\item{output}{the column that should be used as the y-axis:

\if{html}{\out{<div class="sourceCode">}}\preformatted{          options are:
          `num_pts` (raw patient count),
          `prop_retained_start` (proportion patients retained from starting step),
          `prop_retained_prior` (proportion patients retained from prior step),
          `prop_diff_prior` (proportion difference between each step and the prior step)
}\if{html}{\out{</div>}}}
}
\value{
a line graph with the output value for each step and an accompanying
table with the full descriptions of each step
}
\description{
Cohort Attrition \emph{Multi Site Exploratory No Time}
}
