% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_output.R
\name{ca_output}
\alias{ca_output}
\title{Cohort Attrition -- Output Generation}
\usage{
ca_output(
  process_output,
  log_scale = FALSE,
  var_col = "num_pts",
  large_n = FALSE,
  large_n_sites = NULL
)
}
\arguments{
\item{process_output}{\emph{tabular input} || \strong{required}

The tabular output produced by \code{ca_process}}

\item{log_scale}{\emph{boolean} || default to \code{FALSE}

A boolean indicating whether a log transformation should be applied to the
y-axis of the output}

\item{var_col}{\emph{string} || defaults to \code{num_pts}

The name of the column that should be displayed on the plot for Exploratory analyses.
The options are:
\itemize{
\item \code{num_pts}: raw patient count
\item \code{prop_retained_start}: proportion patients retained from the starting step, as indicated by \code{start_step_num}
\item \code{prop_retained_prior}: proportion patients retained from prior step
\item \code{prop_diff_prior}: proportion difference between each step and the prior step
}}

\item{large_n}{\emph{boolean} || defaults to \code{FALSE}

For Multi-Site analyses, a boolean indicating whether the large N
visualization, intended for a high volume of sites, should be used. This
visualization will produce high level summaries across all sites, with an
option to add specific site comparators via the \code{large_n_sites} parameter.}

\item{large_n_sites}{\emph{vector} || defaults to \code{NULL}

When \code{large_n = TRUE}, a vector of site names that can add site-level information
to the plot for comparison across the high level summary information.}
}
\value{
This function will produce a graph to visualize the results
from \code{ca_process} based on the parameters provided. The default
output is typically a static ggplot or gt object, but interactive
elements can be activated by passing the plot through \code{make_interactive_squba}.
For a more detailed description of output specific to each check type,
see the PEDSpace metadata repository
}
\description{
Using the tabular output generated by \code{ca_process}, this function will build a graph to
visualize the results. Each function configuration will output a bespoke ggplot. Theming can
be adjusted by the user after the graph has been output using \code{+ theme()}. Most graphs can
also be made interactive using \code{make_interactive_squba()}
}
\examples{

#' Build mock study attrition
sample_attrition <- dplyr::tibble('site' = c('Site A', 'Site A', 'Site A', 'Site A'),
                                  'step_number' = c(1,2,3,4),
                                  'attrition_step' = c('step 1', 'step 2', 'step 3', 'step 4'),
                                  'num_pts' = c(100, 90, 70, 50))

#' Execute `ca_process` function
#' This example will use the single site, exploratory, cross sectional
#' configuration
ca_process_example <- ca_process(attrition_tbl = sample_attrition,
                                 multi_or_single_site = 'single',
                                 anomaly_or_exploratory = 'exploratory',
                                 start_step_num = 1) \%>\%
  suppressMessages()

ca_process_example

#' Execute `ca_output` function
ca_output_example <- ca_output(process_output = ca_process_example,
                               log_scale = FALSE,
                               var_col = 'prop_retained_start')

ca_output_example[[1]]
ca_output_example[[2]]

#' Easily convert the graph into an interactive ggiraph or plotly object with
#' `make_interactive_squba()`

make_interactive_squba(ca_output_example[[1]])
}
