% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_process.R
\name{ca_process}
\alias{ca_process}
\title{Cohort Attrition}
\usage{
ca_process(
  attrition_tbl,
  multi_or_single_site = "single",
  anomaly_or_exploratory = "exploratory",
  start_step_num = 0,
  var_col = "num_pts",
  p_value = 0.9
)
}
\arguments{
\item{attrition_tbl}{\emph{tabular input} || \strong{required}

A table or CSV file with attrition information for each site included in the cohort.
This table should minimally contain:
\itemize{
\item \code{site} | \emph{character} | the name of the institution
\item \code{step_number} | \emph{integer} | a numeric identifier for the attrition step
\item \code{attrition_step} | \emph{character} | a description of the attrition step
\item \code{num_pts} | \emph{integer} | the patient count for the attrition step
}}

\item{multi_or_single_site}{\emph{string} || defaults to \code{single}

A string, either \code{single} or \code{multi}, indicating whether a single-site or
multi-site analysis should be executed}

\item{anomaly_or_exploratory}{\emph{string} || defaults to \code{exploratory}

A string, either \code{anomaly} or \code{exploratory}, indicating what type of results
should be produced.

Exploratory analyses give a high level summary of the data to examine the
fact representation within the cohort. Anomaly detection analyses are
specialized to identify outliers within the cohort.}

\item{start_step_num}{\emph{integer} || defaults to \code{0}

The \code{step_number} from the \code{attrition_tbl} that should be considered the "start" for
the analysis. This will drive comparisons of later steps to the start.}

\item{var_col}{\emph{string} || defaults to \code{num_pts}

The name of the column that should be used to conduct the analysis for
the Multi-Site, Anomaly Detection, Cross-Sectional check. The options are:
\itemize{
\item \code{num_pts}: raw patient count
\item \code{prop_retained_start}: proportion patients retained from the starting step, as indicated by \code{start_step_num}
\item \code{prop_retained_prior}: proportion patients retained from prior step
\item \code{prop_diff_prior}: proportion difference between each step and the prior step
}}

\item{p_value}{\emph{numeric} || defaults to \code{0.9}

The p value to be used as a threshold in the Multi-Site,
Anomaly Detection, Cross-Sectional analysis}
}
\value{
This function will return a dataframe with all the original attrition
information, plus columns examining the difference between each step and others.
If the Multi-Site, Anomaly Detection, Cross-Sectional check is run, this output
will also include some descriptive statistics about the chosen \code{var_col} and
an indication of which sites are outliers at each attrition step.
}
\description{
This is a plausibility module that will assess attrition counts and how counts
change from each step to the next to help identify the most impactful and any
potentially anomalous attrition criteria. The user will provide a precomputed
attrition. A sample \code{attrition_tbl} with the expected formatting can be found
using \verb{cohortattrition::}.
}
\examples{

#' Build mock study attrition
sample_attrition <- dplyr::tibble('site' = c('Site A', 'Site A', 'Site A', 'Site A'),
                                  'step_number' = c(1,2,3,4),
                                  'attrition_step' = c('step 1', 'step 2', 'step 3', 'step 4'),
                                  'num_pts' = c(100, 90, 70, 50))

#' Execute `ca_process` function
#' This example will use the single site, exploratory, cross sectional
#' configuration
ca_process_example <- ca_process(attrition_tbl = sample_attrition,
                                 multi_or_single_site = 'single',
                                 anomaly_or_exploratory = 'exploratory',
                                 start_step_num = 1)

ca_process_example

#' Execute `ca_output` function
ca_output_example <- ca_output(process_output = ca_process_example,
                               log_scale = FALSE,
                               var_col = 'prop_retained_start')

ca_output_example

#' Easily convert the graph into an interactive ggiraph or plotly object with
#' `make_interactive_squba()`

make_interactive_squba(ca_output_example[[1]])
}
