% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_attrition}
\alias{sample_attrition}
\title{Sample Attrition File}
\format{
\subsection{\code{sample_attrition}}{

A data frame with 10 rows and 4 columns
\describe{
\item{site}{The name of the institution with which the attrition counts are associated}
\item{step_number}{An integer indicating the attrition step number}
\item{attrition_step}{A string describing the attrition step}
\item{num_pts}{The patient count}
}
}
}
\usage{
sample_attrition
}
\description{
A sample version of the file structure expected for the attrition_table
parameter in the \code{ca_process} function. The user should recreate this file
structure and include their study-specific attrition information.
}
\keyword{datasets}
