% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_output.R
\name{ca_output}
\alias{ca_output}
\title{Cohort Attrition Output}
\usage{
ca_output(
  process_output,
  output_function,
  log_scale = FALSE,
  var_col = "num_pts"
)
}
\arguments{
\item{process_output}{the output of the \code{ca_process} function}

\item{output_function}{the name of the output function that should be used provided in the \code{parameter_summary} csv
file that is output to the provided results folder after running the \code{ca_process} function}

\item{log_scale}{a logical indicating whether the results should be shown in a log scale}

\item{var_col}{the column of the output variable of interest. options are:
\code{num_pts} -- the raw count of patients meeting the requirements for each step
\code{prop_retained_start} -- the proportion of patients retained at each step compared to the user-selected starting step
\code{prop_retained_prior} -- the proportion of patients retained at each step compared to the prior step
\code{prop_diff_prior} -- the proportion difference between each step and the step prior}
}
\value{
for ss_exp_nt & ms_exp_nt, a line graph displaying the var_col of interest at each
attrition step is returned, along with a table with the descriptors for each step

\if{html}{\out{<div class="sourceCode">}}\preformatted{    for ms_anom_nt, a dot plot is returned where anomalous values are shown as stars.
    the size of the dot represents the mean value, while the color represents the value
    of the output column
}\if{html}{\out{</div>}}
}
\description{
Cohort Attrition Output
}
