% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_process.R
\name{ca_process}
\alias{ca_process}
\title{Cohort Attrition}
\usage{
ca_process(
  attrition_tbl,
  multi_or_single_site,
  anomaly_or_exploratory,
  start_step_num = 0,
  var_col = "num_pts",
  p_value = 0.9
)
}
\arguments{
\item{attrition_tbl}{table with attrition information for each site needed for analysis
should have at least the following columns:
- @site
- @step_number
- @attrition_step
- @num_pts}

\item{multi_or_single_site}{Option to run the function on a single vs multiple sites
- @single - run the function for a single site
- @multi - run the function for multiple sites}

\item{anomaly_or_exploratory}{Option to conduct an exploratory or anomaly detection analysis. Exploratory analyses give a high
level summary of the data to examine the fact representation within the cohort. Anomaly detection
analyses are specialized to identify outliers within the cohort.}

\item{start_step_num}{the @step_number that should be considered the "start" for the analysis}

\item{var_col}{the column that should be used to conduct the analysis for multi-site anomaly detection. options are:
- @num_pts (raw patient count),
- @prop_retained_start (proportion patients retained from starting step),
- @prop_retained_prior (proportion patients retained from prior step),
- @prop_diff_prior (proportion difference between each step and the prior step)}

\item{p_value}{the p value to be used as a threshold in the multi-site anomaly detection analysis}
}
\value{
a CSV file with all the original attrition information, plus columns examining the difference between each step and others.

\if{html}{\out{<div class="sourceCode">}}\preformatted{    if a multi-site anomaly detection analysis is run, this output will also include some descriptive statistics about the chosen
    `var_col` and an indication of which sites are outliers at each attrition step
}\if{html}{\out{</div>}}
}
\description{
Cohort Attrition
}
