% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_process.R
\name{ca_process}
\alias{ca_process}
\title{Cohort Attrition}
\usage{
ca_process(
  attrition_tbl,
  multi_or_single_site,
  anomaly_or_exploratory,
  start_step_num = 0,
  var_col = "num_pts",
  p_value = 0.9
)
}
\arguments{
\item{attrition_tbl}{table with attrition information for each site needed for analysis
should have at least the following columns:
\itemize{
\item \code{site}
\item \code{step_number}
\item \code{attrition_step}
\item \code{num_pts}
}}

\item{multi_or_single_site}{Option to run the function on a single vs multiple sites
- \code{single} - run the function for a single site
- \code{multi} - run the function for multiple sites}

\item{anomaly_or_exploratory}{Option to conduct an exploratory or anomaly detection analysis. Exploratory analyses give a high
level summary of the data to examine the fact representation within the cohort. Anomaly detection
analyses are specialized to identify outliers within the cohort.}

\item{start_step_num}{the \code{step_number} that should be considered the "start" for the analysis; defaults to 0}

\item{var_col}{the column that should be used to conduct the analysis for multi-site anomaly detection. options are:
- \code{num_pts}: raw patient count
- \code{prop_retained_start}: proportion patients retained from starting step
- \code{prop_retained_prior}: proportion patients retained from prior step
- \code{prop_diff_prior}: proportion difference between each step and the prior step}

\item{p_value}{the p value to be used as a threshold in the multi-site anomaly detection analysis}
}
\value{
a dataframe with all the original attrition information, plus columns examining the difference between each step and others.

\if{html}{\out{<div class="sourceCode">}}\preformatted{    if a multi-site anomaly detection analysis is run, this output will also include some descriptive statistics about the chosen
    `var_col` and an indication of which sites are outliers at each attrition step
}\if{html}{\out{</div>}}
}
\description{
This is a plausibility module that will assess attrition counts and how counts
change from each step to the next to help identify the most impactful and any
potentially anomalous attrition criteria. The user will provide a precomputed
attrition. A sample \code{attrition_tbl} with the expected formatting can be found
using \verb{cohortattrition::}.
}
\examples{

#' Build mock study attrition
sample_attrition <- dplyr::tibble('site' = c('Site A', 'Site A', 'Site A', 'Site A'),
                                  'step_number' = c(1,2,3,4),
                                  'attrition_step' = c('step 1', 'step 2', 'step 3', 'step 4'),
                                  'num_pts' = c(100, 90, 70, 50))

#' Execute `ca_process` function
#' This example will use the single site, exploratory, cross sectional
#' configuration
ca_process_example <- ca_process(attrition_tbl = sample_attrition,
                                 multi_or_single_site = 'single',
                                 anomaly_or_exploratory = 'exploratory',
                                 start_step_num = 1)

ca_process_example

#' Execute `ca_output` function
ca_output_example <- ca_output(process_output = ca_process_example,
                               output_function = 'ca_ss_exp_cs',
                               log_scale = FALSE,
                               var_col = 'prop_retained_start')

ca_output_example

#' Easily convert the graph into an interactive ggiraph or plotly object with
#' `make_interactive_squba()`

make_interactive_squba(ca_output_example[[1]])
}
