% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_process_subfuncs.R
\name{compute_attrition_diff}
\alias{compute_attrition_diff}
\title{Compute difference between attrition steps}
\usage{
compute_attrition_diff(attrition_tbl, start_step_num = 0, site_col = "site")
}
\arguments{
\item{attrition_tbl}{CSV file or dataframe with attrition information -
should include the following columns:
\itemize{
\item \code{num_pts}
\item \code{step_number}
\item \code{attrition_step}
\item \code{site}
}}

\item{start_step_num}{integer indicating the number of the "start" step against which other steps should
be compared; defaults to 0}

\item{site_col}{the column in the attrition table with the name(s) of the site(s)}
}
\value{
the attrition information plus columns that describe the patient drop & percent difference between
each step and between each step and step 0
}
\description{
Compute difference between attrition steps
}
